% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example.r
\name{example}
\alias{example}
\title{Example  of gene-based scoring for SCZ GWAS on chromosome 18.}
\usage{
example()
}
\description{
This function wraps the example of a SNP analysis on chromosome 18.
}
\details{
example.genome.url = system.file("data/ENSEMBL66_gene_bld37_example.txt", package="LDsnpR")   example.snpdata.url = system.file("data/pgc2_scz_chr18_example.txt", package="LDsnpR"); example.ld.file  = system.file("data/chr18_LD1kgEUR.h5", package="LDsnpR")
}

